Always show actual literals in bytecode printer
Currently the keyword "literal" is used inconsistently across the bytecode
instructions. For example, push_const uses it, but set_const doesn't.

Always use "literal" to indicate the index to literals. And this enables
Rubinius' built-in bytecode printer to show the actual literals referenced by
the index.

This would really ease my life when hacking on Rubinus, hopefully others
starting to hack on it.