Introduce fast path for nested constants
We didn't have caching for constant lookup for nested constants, such as
Rubinius::VariableScope. This means that when a constants was referenced
in that way, it would not cache the VariableScope constant.

This commit introduces a constant cache very similar to not nested
constants. It also stores the scope it was used under, so we can
properly validate the have the correct constant. Also specialization is
added to the JIT to have a fast path for these cached constants if they
are still valid.