Add symbols as resources to Process.(g|s)et_rlimit
As of Ruby 1.9 you can pass a symbol as the name of the resource you're
operating on instead of using the constant form.  For example:

    Process.getrlimit(:CORE)
    => [0, 9223372036854775807]

whereas traditionally one would have used:

    Process.getrlimit(Process::RLIMIT_CORE)
    => [0, 9223372036854775807]

This commit adds behaviour to Process.get_rlimit and Process.set_rlimit
to infer and look up a constant name from a resource name supplied as a
Symbol.