Don't inline calls for not often called methods
This changes the JIT so we don't inline methods that aren't often
called, allowing for more room in JIT'ted code for methods that are
often called.

It also changes to keep track of compiling explicitly instead of
(ab)using the call_count for that. Resetting the call count had other
effects such as disabling that method for future jitting that has
adverse effects.

On my system this shaves off around 1 - 2 seconds of a CI run.