Add better version of kind_of optimization
This makes sure to only do a quick check against VM types since if then
the types don't match, we know the whole chain doesn't match. With the
older implementation there was a problem with subclasses of VM types,
because they share the type info but can't use a quick check.

That is why we now compare against the class_id and not use the TypeInfo
directly.