Use WeakRef to track inheritance for serial increments
This sets up weakrefs for each module where we track modules or classes
that this module is included into or that this class is a parent of.
This allows us to increase the serial when a method is added to a
certain module or class. In that case we bump the serial of all the
classes below in the hierarchy so they get invalidated as well.