Use class_id for inline caches and add serial property
With this change we keep track of a serial per class and the combination
of the class_id and the serial_id is used to invalidate inline caches.
This means that we can remove the inline cache registry in the future
and do inline cache invalidation pure on these serials.

This is pretty much the same strategy as JRuby uses. It trades
invaliding caches based on classes for on name. This means that we
should have a lot less invalidations and we don't need the relatively
expensive operation of removing all caches with the same name.