Pass ldflags after object files
When we are finally statically linking bin/vm, we must be careful the order of
arguments to the linker. Gnnerally speaking, we shouldn't pass ldflags first
before object files because the order matters with the static linking.

To be specifc, we can't find library symbols used in the object files in this
order.

So, just reverse the order of arguments.

Previously, there was no problem because LLVM din't depend on any other library
symbols. But when OProfile is enabled this is not the case.