Add Rubinius specific spec for string building
This ensures that we cover the behavior of string interpolation where
the string doesn't have an encoding yet. This is possible on Rubinius,
we only setup the encoding then in the method retrieving the encoding
from the string.

This is an extraction from issue #2222 and is added to prevent
regressions in this behavior.