Store pointer to ConstantCache directly at site
This allows for having a fixed point in memory that the JIT can use in
the future, so we don't have to pin ConstantCache and we can fold the
ConstantCache and ConstantCacheEntry together, while still being thread
safe.

This is the same mechanism that the InlineCache uses by replacing the
symbol that gives the name for the lookup.