Fix variable scope race condition
This assertion could result in being hit due to a race condition. This
would happen if we were actually in the middle of flushing to the heap
while another thread reads variables from this scope.

In this situation, it's actually safe to just read the variable, since
the other thread is flushing it so the variable is still the same.
We already know that isolated_ will always be set before the stack goes
out of scope, so in this situation we don't end up reading invalid data.