Switch to using a vector for tracking locked objects
This improves performance on a lock / unlock benchmark by a factor of
2x. The problem is that std::list allocates memory for every list entry,
so this creates significant overhead if we add and remove linked list
elements often. Switching to a vector removes this overhead, since it
doesn't need to allocate memory on each addition.

We also add a small optimization for which element to remove. Often
locks are used in a stack fashion in code, so often the last element
will be removed in the list because that was also added last. We
optimize for this case so we only have to do the linear search when it's
not the last element.