Use naming conventions the same to other places constructing call frames
We call the existing call frame previous, so we do that here as well, so
when for example doing the check_async, we pass down the new call frame
instead of the old one. This could cause issues, since that call frame
doesn't refer properly to the objects that the current frame has handles
to.

This name change wasn't directly necessary, but doing this makes it
consistent so issues aren't overlooked easily like they were now causing
issues.