Properly setup scope when doing uncommon exit and inlined blocks
This code had a hardcoded assumption that there was only one inlined
block, but there can of course be multiple. This code therefore walks up
the call chain to find the parent machine code to it can setup the scope
in the proper place.

This was causing issues where in case of an uncommon exit and multiple
inlined blocks, it would read bogus locals from another call frame.