Switch to using vectors for tracking JIT runtime data
Vector does suffer from having to grow from time to time, but this is
less problematic than using std::list in these cases because they
require an allocation for each element added. We also don't need any of
the different semantics here of list, so using a vector works better.