Use different way of storing call site information
When we update the call site with a new object object, we need to be
able to through the write barrier for storing the new call site. The
easiest way is to store the compiled code object and the instruction
pointer, so we can use that for storing a new object on the machine code
object.