accept a Symbol as the second argument to Signal.trap(…).
On 1.9 mode Signal.trap can take a Symbol as a second argument in place of a
String. For example, you can say:

  Signal.trap :SIGUSR1, :DEFAULT
  Signal.trap :SIGUSR1, :SIG_DFT
  Signal.trap :SIGUSR1, :SIG_IGN
  Signal.trap :SIGUSR1, :IGNORE

I found this while trying to use Foreman on Rubinius. It has code that uses
:DEFAULT in place of "DEFAULT".

Side-note:
MRI 1.9 & 1.8 behavior is to accept any string or symbol as the second argument.
Rubinius(1.8 & 1.9) mode raise when the second argument isn't something it
understands. This is valid on CRuby, but not rubinius:

  Signal.trap :SIGUSR1, "ninja"