Generate global symbol names for primitives
Before the primitive hookup logic actually took around 0.2s on my system
of the startup time. This was due to how the algorithm works, basically
resulting in number of primitives squared symbol lookups.

With generating these names like this, we remove the string to symbol
lookup needed, resulting in 0.2s faster startup on my system.