Use more platform agnostic mechanism for offset detection
The size of ByteArray is dependent on platform alignment, which is
different for example on 32 bit ARM systems (sizeof(ByteArray) is 24
bytes there, not 20 like 32 bit X86 systems).

By using the actual runtime structure to determine the byte offset, we
can make sure that we actually have a correct value on other platforms
as well.