Switch to primitives for truncate/ftruncate
Sadly this is for now by far the easiest solution for this. The problem
is that on Linux, truncate binds to either truncate or truncate64
depending on compile time flags.

We try to make sure to use 64 bit filesystem operations so we can
support large files properly, but that means that when binding with FFI
it can bind to the wrong truncate function. This problem is especially
apparent on 32 bit ARM, where it sees a wrong offset value, creating
huge files when the specs try to truncate them.

Moving these to primitives make sure they are bound to the proper
functions at compile time. We already do this for other functions taking
an off_t parameter, such as seek to prevent this issue.