Use spinlocks over object lock in method table
We need to lock around reads of the method table as well, and with
object locks we need access to the call frame because it might GC in the
mean while. Changing this is tricky, since it also affects for example
the JIT which doesn't have these structures.

Therefore this switches it to a spinlock, because there are no GC paths
necessary in the method table and this way we can lock around it
properly for reads as well.