Use faster handle validation mechanism
The handle allocator already has a faster way of validating whether a
handle is valid, so we should use that instead of iterating all handles.

This improves performance for JSON entries, since they use
rb_gc_mark_maybe extensively. In a Sidekiq demo app, this validation
took around 3% of the total runtime, with this change it was reduced to
almost nothing.