Store include dirs in cflags and ldflags
This makes passing down custom options easier, since on some systems we
want to add -Wl,rpath= options for linking but they aren't supported
everywhere. Instead of having to duplicate platform detection for that
everywhere, we do it once and store in the cflags and ldflags.

Also removes linking crypto and ssl on OpenBSD/NetBSD since it's not
needed there for the vm itself. Also make sure to pass down cflags and
ldflags properly to extension building.