Use a spinlock around variable scope retrieval
This is needed because these scopes might be used in different threads,
the simplest example being the following case:

def foo
  a = 1
  Thread.new do
    p a
  end
end

Here the local a is accessed from a different thread. This presents a
problem with flushing the scope to the heap when the foo method returns.
In that case the locals aren't on the stack anymore, but the other
thread might still read the data from the stack due to a race condition.

New threads isn't the only case where this might happen, it can also be
for example the scope capture with a proc or lambda and passed to
another thread for execution. If that accesses the defining scope, this
presents the same problem.

This fixes one cause of crashes of the reproduction in issue #2339.