Grabbing the return value shouldn't clear exception state
This was unconditionally removing the thread state when retrieving this
value, but it only makes sense when the state is either cBreak or
cReturn, since this code is always used on conjunction with
rbx_return_to_here or rbx_break_to_here.

This could cause a bug in the interpreter control logic, since now a
method could have returned a NULL and set the exception state, but this
would clear the exception state. This would result in hitting the
rubinius::bug assertion in the interpreter at instructions.cpp:202,
aborting with the "Control flow error in interpreter" error.

Fixes another bug when I was running the repro from #2339.