Add additional GC debugging that is disabled by default
This is some very basic GC debugging that I've been using. It basically
scans the current call frames after a GC to see if the values referenced
there and on the interpreter scan are updated properly.

This was mainly used to debug world halting issues before, still had it
around and thought it would be useful to add so other can enable this as
well for certain cases. The overhead is fairly minimal, only adding a
small time to GC time and no overhead when not GC'ing.