Setup lock around tracking of fiber data for stacks
We now have a specialized finalizer thread that also runs Fiber
finalizers, unlike in the past where the GC would directly clean this
up. This does however introduce a race condition here for the data
structure we use to track the fiber data's.

The finalizer thread and another thread could be adding a new fiber data
and removing it at the same time. This could corrupt the datas_ set of
active fiber data. Therefore we should lock around it.

Also found when running #2339.