Move Thread setup into the VM and fix GC issues with rb_thread_create
This moves thread setup into the VM, because if it happens in Ruby it
might trigger a GC. This is problematic with rb_thread_create, which
uses a void* pointer to pass arguments around that might get cleaned up
too soon then.

By using setup this way, we can ensure no GC happens before the new
thread is properly started and has extracted the possible argument. We
also need to make sure we reload the thread object in case of an
exception so we don't accidentally refer to an old / moved version of
the object.