Wrap up strerror_r usage inside macro
This wraps up the uglyness needed to hide the annoying fact that
strerror_r has different signatures for GNU and XSI. On GNU it also
marks the return value as not ignorable. By using a hacky if() around
it, we can ignore the value.

We wrap up the uglyness in a macro so we don't see it everywhere...

Fixes #2363